@@ -1,11 +1,10 @@
-﻿namespace MAUIDnr1.Models
+﻿namespace MAUIDnr1.Models;
+
+public class File
 {
-    public class File
-    {
-        public int Id { get; set; }
-        public int ShowId { get; set; }
-        public string Url { get; set; }
-        public int Minutes { get; set; }
-        public int Bytes { get; set; }
-    }
+    public int Id { get; set; }
+    public int ShowId { get; set; }
+    public string Url { get; set; }
+    public int Minutes { get; set; }
+    public int Bytes { get; set; }
 }
