@@ -1,13 +1,12 @@
-﻿namespace MAUIDnr1.Models
+﻿namespace MAUIDnr1.Models;
+
+public class Sponsor
 {
-    public class Sponsor
-    {
-        public int Id { get; set; }
-        public string Name { get; set; } = "";
-        public string BannerUrl { get; set; } = "";
-        public string Blurb { get; set; } = "";
-        public string LandingUrl { get; set; } = "";
-        public string Copy { get; set; } = "";
-        public string LastError { get; set; } = "";
-    }
+    public int Id { get; set; }
+    public string Name { get; set; } = "";
+    public string BannerUrl { get; set; } = "";
+    public string Blurb { get; set; } = "";
+    public string LandingUrl { get; set; } = "";
+    public string Copy { get; set; } = "";
+    public string LastError { get; set; } = "";
 }
