@@ -1,11 +1,14 @@
 ﻿@page "/details/{ShowNumber}"
 
 <span @onclick="NavigateHome" class="oi oi-arrow-left"
-      style="zoom:2;" aria-hidden="true"></span>
+      style="zoom:2;cursor:pointer;" aria-hidden="true"></span>
 
 @if (PlayingPlayList)
 {
-    <h3>Playlist: @AppState.SelectedPlayList.Name (@PlaylistShowIndex of @PlaylistShowCount)</h3>
+    <h3>
+        Playlist: @AppState.SelectedPlayList.Name
+        (@PlaylistShowIndex of @PlaylistShowCount)
+    </h3>
 }
 
 @if (ThisShow != null)
@@ -56,7 +59,8 @@
         {
             <h3>@guest.Name</h3>
             <div style="padding:20px;">
-                <img src="@guest.PhotoUrl" style="width:160px;padding-right:10px;padding-bottom:10px;float:left;" />
+                <img src="@guest.PhotoUrl"
+         style="width:160px;padding-right:10px;padding-bottom:10px;float:left;" />
                 <div>@guest.Bio</div>
             </div>
             <br clear="left" />
@@ -69,7 +73,8 @@
         <div style="padding-left:20px;">
             @foreach (var link in ThisShow.ShowDetails.Links)
             {
-                <a style="padding-bottom:10px;" target="_blank" href="@link.Url">@link.Text</a>
+                <a style="padding-bottom:10px;"
+       target="_blank" href="@link.Url">@link.Text</a>
                 <br />
             }
         </div>
